+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Update description(template)
 ### Fixed
 *  -->
 
+## [0.8.6] - 2022-04-07
+### Changed
+* Simplified bindings name for basic and extra keys
+* Moved f keys logic to macrosengine
+### Removed
+* F keys from bindings
 ## [0.8.5] - 2022-03-01
 ### Added
 * SSprintf that functions like printf for logging purposes